Small bug in the voronoi calculation
The code does not account for loop in the voronoi diagram. Recent calculation on smaller badly defined objects (fish) have shown the existence of this loop. This is a quick fix, the logical structure of looking for the biggest branch in the tree should also be modified (bigbranche.m).

Those events are rare, I have never seen that with plants. So there is no urgency to change bigbranche.m.